# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Login - Desafio",
    "version": "1.0.0",
    "description": "API para simulação de login, bloqueio e reset de senha."
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "Realiza login do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": { "type": "string" },
                  "senha": { "type": "string" }
                },
                "required": ["usuario", "senha"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas, senha incorreta ou bloqueada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/resetar-senha": {
      "post": {
        "summary": "Reseta a senha do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": { "type": "string" },
                  "novaSenha": { "type": "string" }
                },
                "required": ["usuario", "novaSenha"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha resetada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
} 